fixes to cluster configuration for multicluster

+++ clc/modules/www/src/main/java/edu/ucsb/eucalyptus/admin/client/SystemConfigTable.java
@@ -175,7 +175,7 @@ public class SystemConfigTable extends VerticalPanel {
 		if(walrusList.size() > 0)
 			walrusInfo = walrusList.get(0);
 		else 
-			walrusInfo = new WalrusInfoWeb("walrus-name", "hi", 5, 5120, 50000, 50000);
+			walrusInfo = new WalrusInfoWeb("walrus-name", "hi", 5l, 5120l, 50000l, 50000l);
 
 		this.w_grid.clear ();
 		this.w_grid.resize ( 4, 2 );
@@ -286,10 +286,10 @@ public class SystemConfigTable extends VerticalPanel {
 		WalrusInfoWeb walrusInfo = walrusList.get(0);
 		this.SystemConfig.setWalrusUrl                (this.walrusURL_box.getText());
 		walrusInfo.setBucketsRootDirectory               (this.walrusPath_box.getText());
-		walrusInfo.setMaxBucketsPerUser  (Integer.parseInt(this.maxBuckets_box.getText()));
-		walrusInfo.setMaxBucketSizeInMB  (Integer.parseInt(this.maxBucketSize_box.getText()));
-		walrusInfo.setMaxCacheSizeInMB   (Integer.parseInt(this.maxCacheSize_box.getText()));
-		walrusInfo.setSnapshotsTotalInGB (Integer.parseInt(this.totalSnapshots_box.getText()));
+		walrusInfo.setMaxBucketsPerUser  (Long.parseLong(this.maxBuckets_box.getText()));
+		walrusInfo.setMaxBucketSizeInMB  (Long.parseLong(this.maxBucketSize_box.getText()));
+		walrusInfo.setMaxCacheSizeInMB   (Long.parseLong(this.maxCacheSize_box.getText()));
+		walrusInfo.setSnapshotsTotalInGB (Long.parseLong(this.totalSnapshots_box.getText()));
 		this.SystemConfig.setDefaultKernelId           (this.defaultKernel_box.getText());
 		this.SystemConfig.setDefaultRamdiskId          (this.defaultRamdisk_box.getText());
 	}
